Change to functional/class
