+++ app/javascript/packs/application.js
@@ -16,6 +16,6 @@ require('channels');
 // const images = require.context('../images', true)
 // const imagePath = (name) => images(name, true)
 // Support component names relative to this directory:
-const containerRequireContext = require.context('containers', true);
+const containerRequireContext = require.context('components', true);
 const ReactRailsUJS = require('react_ujs');
 ReactRailsUJS.useContext(containerRequireContext);
