+++ internals/generators/class/index.js
@@ -1,25 +1,21 @@
 /**
- * Component Generator
+ * Class Component Generator
  */
-
-/* eslint strict: ["off"] */
-
-'use strict';
 
 const componentExists = require('../utils/componentExists');
 
 module.exports = {
-  description: 'Add an unconnected component',
+  description: 'Add a class component',
   prompts: [
     {
       type: 'input',
       name: 'name',
       message: 'What should it be called?',
-      default: 'Button',
+      default: 'Form',
       validate: value => {
         if (/.+/.test(value)) {
           return componentExists(value)
-            ? 'A component or container with this name already exists'
+            ? 'A component with this name already exists'
             : true;
         }
 
@@ -28,9 +24,9 @@ module.exports = {
     },
     {
       type: 'confirm',
-      name: 'memo',
+      name: 'wantHeaders',
       default: false,
-      message: 'Do you want to wrap your component in React.memo?',
+      message: 'Do you want headers?',
     },
     {
       type: 'confirm',
@@ -38,17 +34,11 @@ module.exports = {
       default: true,
       message: 'Do you want styles?',
     },
-    // {
-    //   type: 'confirm',
-    //   name: 'wantMessages',
-    //   default: true,
-    //   message: 'Do you want i18n messages (i.e. will this component use text)?',
-    // },
     {
       type: 'confirm',
       name: 'wantLoadable',
-      default: false,
-      message: 'Do you want to load the component asynchronously?',
+      default: true,
+      message: 'Do you want to load resources asynchronously?',
     },
   ],
   actions: data => {
@@ -57,42 +47,26 @@ module.exports = {
       {
         type: 'add',
         path: '../../app/javascript/components/{{properCase name}}/index.js',
-        templateFile: './component/index.js.hbs',
-        abortOnFail: true,
-      },
-      {
-        type: 'add',
-        path: '../../app/javascript/components/{{properCase name}}/tests/index.test.js',
-        templateFile: './component/test.js.hbs',
+        templateFile: './class/index.js.hbs',
         abortOnFail: true,
       },
     ];
 
-    // If user wants styles
+    // If component wants styles
     if (data.wantStyles) {
       actions.push({
         type: 'add',
         path: '../../app/javascript/components/{{properCase name}}/styles.js',
-        templateFile: './component/styles.js.hbs',
+        templateFile: './class/styles.js.hbs',
         abortOnFail: true,
       });
     }
-    // If the user wants i18n messages
-    // if (data.wantMessages) {
-    //   actions.push({
-    //     type: 'add',
-    //     path: '../../app/javascript/components/{{properCase name}}/messages.js',
-    //     templateFile: './component/messages.js.hbs',
-    //     abortOnFail: true,
-    //   });
-    // }
 
-    // If the user wants Loadable.js to load the component asynchronously
     if (data.wantLoadable) {
       actions.push({
         type: 'add',
         path: '../../app/javascript/components/{{properCase name}}/Loadable.js',
-        templateFile: './component/loadable.js.hbs',
+        templateFile: './functional/loadable.js.hbs',
         abortOnFail: true,
       });
     }
