+++ internals/generators/class/styles.js.hbs
