--- internals/generators/container/messages.js.hbs
@@ -1,16 +1,0 @@
-/*
- * {{ properCase name }} Messages
- *
- * This contains all the text for the {{ properCase name }} container.
- */
-
-import { defineMessages } from 'react-intl';
-
-export const scope = 'app.containers.{{ properCase name }}';
-
-export default defineMessages({
-  header: {
-    id: `${scope}.header`,
-    defaultMessage: 'This is the {{ properCase name }} container!',
-  },
-});
