+++ internals/generators/functional/index.js.hbs
@@ -18,23 +18,14 @@ import { compose } from 'redux';
 {{#if wantStyles}}
 import { withStyles } from '@material-ui/core/styles';
 {{/if}}
-{{!-- {{#if wantMessages}}
-import { FormattedMessage } from 'react-intl';
-{{/if}} --}}
 
 {{#if wantStyles}}
 import styles from './styles';
 {{/if}}
-{{!-- {{#if wantMessages}}
-import messages from './messages';
-{{/if}} --}}
 
 function {{ properCase name }}({{#if wantStyles}}{ classes }{{/if}}) {
 return (
 <div{{#if wantStyles}} className={ classes.root }{{/if}}>
-  {{!-- {{#if wantMessages}}
-      <FormattedMessage {...messages.header} />
-    {{/if}} --}}
   </div>
   );
   }
@@ -45,11 +36,4 @@ return (
   {{/if}}
   };
 
-  export default compose(
-  {{#if wantStyles}}
-  withStyles(styles),
-  {{/if}}
-  {{#if memo}}
-  memo,
-  {{/if}}
-  )({{ properCase name }});
+  export default {{#if memo}}memo{{/if}}({{#if wantStyles }}withStyles(styles){{/if}}({{ properCase name }}));
