+++ internals/generators/functional/loadable.js.hbs
