+++ internals/generators/functional/styles.js.hbs
