--- internals/generators/language/add-locale-data.hbs
@@ -1,1 +1,0 @@
-$1addLocaleData({{language}}LocaleData);
