--- internals/generators/language/app-locale.hbs
@@ -1,1 +1,0 @@
-$1  '{{language}}',
