--- internals/generators/language/format-translation-messages.hbs
@@ -1,1 +1,0 @@
-$1  {{language}}: formatTranslationMessages('{{language}}', {{language}}TranslationMessages),
