--- internals/generators/language/intl-locale-data.hbs
@@ -1,1 +1,0 @@
-$&const {{language}}LocaleData = require('react-intl/locale-data/{{language}}');
