--- internals/generators/language/polyfill-intl-locale.hbs
@@ -1,1 +1,0 @@
-$1        import('intl/locale-data/jsonp/{{language}}.js'),
