--- internals/generators/language/translation-messages.hbs
@@ -1,1 +1,0 @@
-$1const {{language}}TranslationMessages = require('./translations/{{language}}.json');
