--- internals/generators/language/translations-json.hbs
@@ -1,1 +1,0 @@
-[]
